requirements.txt
absl-py==1.2.0
black==22.6.0
flake8==5.0.4
tqdm==4.64.0
matplotlib==3.5.3
scipy==1.9.1
rasterio==1.3.4
xgboost
scikit-learn==1.2.0
opencv-python-headless
tensorflow-addons
validators
transunet
evaluate
transformers
